+++ src/components/Passive/Passive.tsx
@@ -4,7 +4,7 @@ export const Passive: React.FunctionComponent = () => {
   
   return(
     <section className="passive">
-      <h1>Passive(in progress)</h1>
+      <h1>Passive<br/>(in progress)</h1>
     </section>
   )
 }
