+++ src/components/Stocks/Stocks.tsx
@@ -4,7 +4,7 @@ export const Stocks: React.FunctionComponent = () => {
   
   return(
     <section className="stocks">
-      <h1>Stocks(in progress)</h1>
+      <h1>Stocks<br/>(in progress)</h1>
     </section>
   )
 }
